# yizyiz/target/classes/openapi/api.yaml
openapi: 3.0.3
info:
  title: Task API
  description: |
    A RESTful API for managing tasks using a contract-first approach.
    This API demonstrates OpenAPI specification, Flyway migrations, and PostgreSQL integration.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

tags:
  - name: Tasks
    description: Task management operations

paths:
  /tasks:
    get:
      tags:
        - Tasks
      summary: Get all tasks
      description: Retrieves a list of all tasks in the system
      operationId: getAllTasks
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - Tasks
      summary: Create a new task
      description: Creates a new task with the provided information
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the task
        title:
          type: string
          description: Title of the task
          example: "Complete project documentation"
        description:
          type: string
          description: Detailed description of the task
          example: "Write comprehensive documentation for the Spring Boot API"
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
      required:
        - id
        - title
        - description
        - completed

    TaskRequest:
      type: object
      properties:
        title:
          type: string
          description: Title of the task
          example: "Complete project documentation"
        description:
          type: string
          description: Detailed description of the task
          example: "Write comprehensive documentation for the Spring Boot API"
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
      required:
        - title
        - description

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the error
      required:
        - status
        - message